moved processResponse to ModbusBuffer
