+++ src/LSST/cRIO/ILC.cpp
@@ -163,60 +163,6 @@ void ILC::simulateResponse(bool simulate) {
     }
 }
 
-void ILC::addResponse(uint8_t func, std::function<void(uint8_t)> action, uint8_t errorResponse,
-                      std::function<void(uint8_t, uint8_t)> errorAction) {
-    _actions[func] = action;
-    _errorActions[errorResponse] =
-            std::pair<uint8_t, std::function<void(uint8_t, uint8_t)>>(func, errorAction);
-}
-
-void ILC::processResponse(uint16_t *response, size_t length) {
-    preProcess();
-
-    setBuffer(response, length);
-
-    while (endOfBuffer() == false) {
-        uint8_t address = read<uint8_t>();
-        uint8_t func = read<uint8_t>();
-
-        // either function response was received, or error response. For error
-        // response, check if the function for which it is used was called.
-        if (_errorActions.find(func) == _errorActions.end()) {
-            checkCommanded(address, func);
-        } else {
-            checkCommanded(address, _errorActions[func].first);
-        }
-
-        try {
-            _actions.at(func)(address);
-        } catch (std::out_of_range &_ex) {
-            try {
-                auto errorAction = _errorActions.at(func);
-                uint8_t exception = read<uint8_t>();
-                checkCRC();
-                if (errorAction.second) {
-                    errorAction.second(address, exception);
-                } else {
-                    throw Exception(address, func, exception);
-                }
-            } catch (std::out_of_range &_ex2) {
-                throw UnknownResponse(address, func);
-            }
-        }
-    }
-
-    postProcess();
-}
-
-ILC::UnknownResponse::UnknownResponse(uint8_t address, uint8_t func)
-        : std::runtime_error(fmt::format("Unknown function {1} (0x{1:02x}) in ILC response for address {0}",
-                                         address, func)) {}
-
-ILC::Exception::Exception(uint8_t address, uint8_t func, uint8_t exception)
-        : std::runtime_error(
-                  fmt::format("ILC Exception {2} (ILC address {0}, ILC response function {1} (0x{1:02x}))",
-                              address, func, exception)) {}
-
 uint8_t ILC::nextBroadcastCounter() {
     _broadcastCounter++;
     if (_broadcastCounter > 15) {
